feat: fix label and unused fragment from feedback
Also refactored key press into test helper, since usually keydown
accompanies keyup
